Performance-2nd submission
The data insert has a little bit small problem not so sure.
The pop-up of "An item with the same key has already been added" is because the performance year is duplicated in Access database under table "performance_evaluation", should we change it or redesign data structure? Others like reward claiming already solved and make sure can run already.